% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedLR.R
\name{adjusted.LR}
\alias{adjusted.LR}
\title{Adjusted Log-Rank Test}
\usage{
adjusted.LR(times, failures, variable, weights = NULL)
}
\arguments{
\item{times}{A numeric vector of survival times.}

\item{failures}{A binary vector indicating event occurrence.}

\item{variable}{A numeric binary variable (0 and 1) indicating group membership.}

\item{weights}{A numeric vector of weights.}
}
\value{
A list containing the test statistic and p-value.
}
\description{
Performs an adjusted log-rank test considering weights.
}
\keyword{internal}
