% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jskm.R
\name{jskm}
\alias{jskm}
\title{Creates a Kaplan-Meier plot: ggkm + cumhaz option}
\usage{
jskm(sfit, table = FALSE, xlabs = "Time-to-event",
  ylabs = "Survival (\%)", xlims = c(0, max(sfit$time)), ylims = c(0, 1),
  ystratalabs = names(sfit$strata), ystrataname = "Strata",
  timeby = signif(max(sfit$time)/7, 1), main = "", pval = FALSE,
  marks = TRUE, shape = 3, legend = TRUE, legendposition = c(0.85, 0.8),
  ci = FALSE, subs = NULL, linecols = "Set1", dashed = FALSE,
  cumhaz = F, ...)
}
\arguments{
\item{timeby}{numeric: control the granularity along the time-axis; defaults to 7 time-points. Default = signif(max(sfit$time)/7, 1)}

\item{main}{plot title}

\item{pval}{logical: add the pvalue to the plot?}

\item{marks}{logical: should censoring marks be added?}

\item{subs}{= NULL,}

\item{cumhaz}{Show cumulaive hazard function, Default: F}

\item{...}{PARAM_DESCRIPTION}

\item{sfit:}{a survfit object}

\item{table:}{logical: Create a table graphic below the K-M plot, indicating at-risk numbers?}

\item{xlabs:}{x-axis label}

\item{ylabs:}{y-axis label}

\item{xlims:}{numeric: list of min and max for x-axis. Default = c(0,max(sfit$time))}

\item{ylims:}{numeric: list of min and max for y-axis. Default = c(0,1)}

\item{ystratalabs:}{character list. A list of names for each strata. Default = names(sfit$strata)}

\item{ystrataname:}{The legend name. Default = "Strata"}

\item{shape:}{what shape should the censoring marks be, default is a vertical line}

\item{legend:}{logical. should a legend be added to the plot?}

\item{legendposition:}{numeric. x, y position of the legend if plotted. Default=c(0.85,0.8)}

\item{ci:}{logical. Should confidence intervals be plotted. Default = FALSE}

\item{linecols:}{Character. Colour brewer pallettes too colour lines. Default ="Set1",}

\item{dashed:}{logical. Should a variety of linetypes be used to identify lines. Default = FALSE}
}
\value{
Plot
}
\description{
Creates a Kaplan-Meier plot with at risk tables below
}
\details{
DETAILS
}
\examples{
 library(survival)
 data(colon)
 fit <- survfit(Surv(time,status)~rx, data=colon)
 jskm(fit, timeby=500)
}
\author{
Jinseob Kim, but heavily modified version of a script created by Michael Way.
\url{https://github.com/michaelway/ggkm/}
I have packaged this function, added functions to namespace and included a range of new parameters.
}
