% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyjskm.R
\name{svyjskm}
\alias{svyjskm}
\title{Creates a Weighted Kaplan-Meier plot - svykm.object in survey package}
\usage{
svyjskm(sfit, xlabs = "Time-to-event", ylabs = "Survival (\%)",
  xlims = NULL, ylims = c(0, 1), ystratalabs = NULL,
  ystrataname = "Strata", timeby = NULL, main = "", pval = FALSE,
  legend = TRUE, legendposition = c(0.85, 0.8), linecols = "Set1",
  dashed = FALSE, cumhaz = F, design = NULL, ...)
}
\arguments{
\item{sfit}{a svykm object}

\item{xlabs}{x-axis label, Default: 'Time-to-event'}

\item{ylabs}{y-axis label.}

\item{xlims}{numeric: list of min and max for x-axis. Default: NULL}

\item{ylims}{numeric: list of min and max for y-axis. Default: c(0, 1)}

\item{ystratalabs}{character list. A list of names for each strata. Default: NULL}

\item{ystrataname}{The legend name. Default: 'Strata'}

\item{timeby}{numeric: control the granularity along the time-axis; defaults to 7 time-points.}

\item{main}{plot title, Default: ''}

\item{pval}{logical: add the pvalue to the plot?, Default: FALSE}

\item{legend}{logical. should a legend be added to the plot? Default: TRUE}

\item{legendposition}{numeric. x, y position of the legend if plotted. Default: c(0.85, 0.8)}

\item{linecols}{Character. Colour brewer pallettes too colour lines. Default: 'Set1'}

\item{dashed}{logical. Should a variety of linetypes be used to identify lines. Default: FALSE}

\item{cumhaz}{Show cumulaive hazard function, Default: F}

\item{design}{Data design for reactive design data , Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
plot
}
\description{
Creates a Weighted Kaplan-Meier plot - svykm.object in survey package
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 data(pbc, package="survival")
 pbc$randomized <- with(pbc, !is.na(trt) & trt>0)
 biasmodel<-glm(randomized~age*edema,data=pbc)
 pbc$randprob<-fitted(biasmodel)
 
 dpbc<-svydesign(id=~1, prob=~randprob, strata=~edema, data=subset(pbc,randomized))
 
 s1 <- svykm(Surv(time,status>0)~sex, design=dpbc)
 svyjskm(s1)
 }
}
}
